http://jira.sakaiproject.org/jira/browse/EVALSYS-559 Trying a whole new approach to this which requires new logic layer method, it appears that any exception from within the transaction which crosses the method boundary is trapped and causes the transaction to fail so we could not use that strategy, I should have know that, now I do anyway
git-svn-id: https://source.sakaiproject.org/contrib/evaluation/trunk@48773 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ tool/src/java/org/sakaiproject/evaluation/tool/LocalTemplateLogic.java
@@ -22,7 +22,6 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.sakaiproject.evaluation.constant.EvalConstants;
 import org.sakaiproject.evaluation.logic.EvalAuthoringService;
-import org.sakaiproject.evaluation.logic.exceptions.InUseException;
 import org.sakaiproject.evaluation.logic.externals.EvalExternalLogic;
 import org.sakaiproject.evaluation.model.EvalItem;
 import org.sakaiproject.evaluation.model.EvalScale;
@@ -172,10 +171,10 @@ public class LocalTemplateLogic {
          Long itemId = templateItem.getItem().getId();
          authoringService.deleteTemplateItem(templateItem.getId(), currentUserId);
          // if this parent is used elsewhere then this will cause exception - EVALSYS-559
-         try {
-            authoringService.deleteItem(itemId, currentUserId);
-         } catch (InUseException e) {
+         if (authoringService.isUsedItem(itemId)) {
             log.info("Cannot remove block parent item ("+itemId+") - item is in use elsewhere");
+         } else {
+            authoringService.deleteItem(itemId, currentUserId);
          }
 
          // modify block children template items
